// ForExample/src/app/models/debitor.model.ts
export interface DebitorModel {
  id: number;
  policyId: number;
  insureTypeId: number;
  policySeries: string;
  policyNumber: string;
  currencyId: number;
  currencyName: string;
  insuredId: string;
  insuredName: string;
  insuredType: 'c' | 'p';
  bankId: string;
  bankName: string;
  benefId: string;
  benefName: string;
  branchCode: string;
  branchName: string;
  compval: string;
  ifrs: string;
  cntrl: string;
  policyDebtSum: number;
  policyDebtSumAzn: number;
  paidSum: number;
  premium: number;
  premiumAzn: number;
  paidSumAzn: number;
  paidEdSumAzn: number;
  paidSumCurr: number;
  policyDate: string;
  expiredDebt30: number;
  expiredDebt60: number;
  expiredDebt90: number;
  expiredDebt180: number;
  expiredDebt181: number;
  instSum15Fw: number;
  instSum30Fw: number;
  instSum60Fw: number;
  instSum90Fw: number;
  instSum180Fw: number;
  instSum360Fw: number;
  instSum1080Fw: number;
  instSum1081Fw: number;
  instSumTotalFw: number;
  expiredDebtTotal: number;
  respPersonName: string;
  createUser: string;
  createUserId: number;
  respPersonId: number;
  firstStartDate: string;
  lastEndDate: string;
  pageId: number;
  order: string;
  DebitorDate?: string;
  limit: number;
  offset: number;
}